Finished working on the missiles and basic building types, cities and bases.
Bunch of logical bug fixes.

bases can shoot missiles, the one that shoots depends on which is the closest to the mouse.

Missiles can explode once destination is reached. the explosion animation is linked to the collision as the radius dynamically adjusts which is used for collision detection.

Created psuedo code for spawning enemy missiles.